refactor: Hash -> Cryptographic 로 변경에 따른 수정
